+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## v1.6.10~rc5 - 2026/??/??
 
+- [DEPS] Updated NGINX version to v1.30.0 for all integrations.
 - [DEPS] Updated Modsecurity version to v3.0.15
 - [DEPS] Updated Mbed TLS version to v4.1.0
 - [DEPS] Updated libinjection version to v4.0.0